[ALPS08691756] arm: config: mt6765: Add mt6765 32 bit config
[Description]
Add SPM/cm_mgr config for MT6765 32 bit

[Test]
Init success

MTK-Commit-Id: 06fd24378b9f9a86896370bdde4574983a671edb

CR-Id: ALPS08691756
Feature: SPM Firmware

+++ drivers/misc/mediatek/spm/common_v0/mtk_spm_fs.c
@@ -175,7 +175,7 @@ static ssize_t show_pwr_ctrl(int id, const struct pwr_ctrl *pwrctrl, char *buf)
 	for (i = 0; i < PW_MAX_COUNT; i++) {
 		p += sprintf(p, "%s = 0x%zx\n",
 				pwr_ctrl_str[i],
-				SMC_CALL(GET_PWR_CTRL_ARGS,
+				(size_t)SMC_CALL(GET_PWR_CTRL_ARGS,
 				id, i, 0));
 	}
 	WARN_ON(p - buf >= PAGE_SIZE);
